Reduce malloc calls

+++ src/include/forward_list.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017-2019 Bailey Thompson
+ * Copyright (c) 2017-2020 Bailey Thompson
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -34,28 +34,28 @@ typedef struct internal_forward_list *forward_list;
 forward_list forward_list_init(size_t data_size);
 
 /* Utility */
-int forward_list_size(forward_list me);
+size_t forward_list_size(forward_list me);
 int forward_list_is_empty(forward_list me);
 void forward_list_copy_to_array(void *arr, forward_list me);
 
 /* Adding */
 int forward_list_add_first(forward_list me, void *data);
-int forward_list_add_at(forward_list me, int index, void *data);
+int forward_list_add_at(forward_list me, size_t index, void *data);
 int forward_list_add_last(forward_list me, void *data);
 
 /* Removing */
 int forward_list_remove_first(forward_list me);
-int forward_list_remove_at(forward_list me, int index);
+int forward_list_remove_at(forward_list me, size_t index);
 int forward_list_remove_last(forward_list me);
 
 /* Setting */
 int forward_list_set_first(forward_list me, void *data);
-int forward_list_set_at(forward_list me, int index, void *data);
+int forward_list_set_at(forward_list me, size_t index, void *data);
 int forward_list_set_last(forward_list me, void *data);
 
 /* Getting */
 int forward_list_get_first(void *data, forward_list me);
-int forward_list_get_at(void *data, forward_list me, int index);
+int forward_list_get_at(void *data, forward_list me, size_t index);
 int forward_list_get_last(void *data, forward_list me);
 
 /* Ending */
